requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
motor>=3.3.0
pydantic>=2.4.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
PyPDF2>=3.0.0
langchain>=0.1.0
langchain-google-genai>=0.0.6
